+++ middlewares/authzCourse.js
@@ -18,6 +18,7 @@ module.exports = function(req, res, next) {
 
         var permissions_course = result.rows[0].permissions_course;
         res.locals.course = result.rows[0].course;
+        res.locals.courses = result.rows[0].courses;
 
         if (permissions_course.course_role == 'None') {
             return next(error.make(403, 'Access denied'));