+++ server.js
@@ -141,6 +141,23 @@ config.fileUploadMaxBytesFormatted = filesize(config.fileUploadMaxBytes, {base: 
 app.post('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/uploads', upload.single('file'));
 app.post('/pl/course_instance/:course_instance_id/instance_question/:instance_question_id', upload.single('file'));
 app.post('/pl/course_instance/:course_instance_id/assessment_instance/:assessment_instance_id', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/question/:question_id', upload.single('file'));
+app.post('/pl/course/:course_id/question/:question_id', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/settings', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/instance_admin/settings', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/course_admin/settings', upload.single('file'));
+app.post('/pl/course/:course_id/course_admin/settings', upload.single('file'));
+app.post('/pl/course/:course_id/course_admin/file_view', upload.single('file'));
+app.post('/pl/course/:course_id/course_admin/file_view/*', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/course_admin/file_view', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/course_admin/file_view/*', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/instance_admin/file_view', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/instance_admin/file_view/*', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/file_view', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/file_view/*', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/question/:question_id/file_view', upload.single('file'));
+app.post('/pl/course_instance/:course_instance_id/instructor/question/:question_id/file_view/*', upload.single('file'));
+
 
 // Limit to 1MB of JSON
 app.use(bodyParser.json({limit: 1024 * 1024}));
@@ -237,7 +254,7 @@ app.use('/pl/course_instance/:course_instance_id', function(req, res, next) {res
 // but before instructor authz so we still get a chance to enforce that.
 app.use(/^\/pl\/course_instance\/[0-9]+\/?$/, function(req, res, _next) {
     if (res.locals.authz_data.has_instructor_view) {
-        res.redirect(res.locals.urlPrefix + '/instructor/assessments');
+        res.redirect(res.locals.urlPrefix + '/instructor/instance_admin/assessments');
     } else {
         res.redirect(res.locals.urlPrefix + '/assessments');
     }
@@ -279,19 +296,21 @@ app.use('/pl/api/v1', require('./api/v1'));
 app.use('/pl/course_instance/:course_instance_id/instructor/effectiveUser', [
     require('./pages/instructorEffectiveUser/instructorEffectiveUser'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/assessments', [
-    require('./pages/instructorAssessments/instructorAssessments'),
-]);
 
 // single assessment
 
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id', [
     require('./middlewares/selectAndAuthzAssessment'),
+    require('./middlewares/selectAssessments'),
 ]);
 app.use(/^(\/pl\/course_instance\/[0-9]+\/instructor\/assessment\/[0-9]+)\/?$/, (req, res, _next) => {
     res.redirect(`${req.params[0]}/questions`);
 });
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id', function(req, res, next) {res.locals.navPage = 'assessment'; next();});
+app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorAssessmentSettings/instructorAssessmentSettings'),
+]);
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/questions', [
     function(req, res, next) {res.locals.navSubPage = 'questions'; next();},
     require('./pages/instructorAssessmentQuestions/instructorAssessmentQuestions'),
@@ -301,11 +320,11 @@ app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessme
     require('./pages/instructorAssessmentAccess/instructorAssessmentAccess'),
 ]);
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/assessment_statistics', [
-    function(req, res, next) {res.locals.navSubPage = 'assessmentStatistics'; next();},
+    function(req, res, next) {res.locals.navSubPage = 'assessment_statistics'; next();},
     require('./pages/instructorAssessmentStatistics/instructorAssessmentStatistics'),
 ]);
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/question_statistics', [
-    function(req, res, next) {res.locals.navSubPage = 'questionStatistics'; next();},
+    function(req, res, next) {res.locals.navSubPage = 'question_statistics'; next();},
     require('./pages/shared/assessmentStatDescriptions'),
     require('./pages/shared/floatFormatters'),
     require('./pages/instructorAssessmentQuestionStatistics/instructorAssessmentQuestionStatistics'),
@@ -326,10 +345,15 @@ app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessme
     function(req, res, next) {res.locals.navSubPage = 'instances'; next();},
     require('./pages/instructorAssessmentInstances/instructorAssessmentInstances'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/edit', [
-    function(req, res, next) {res.locals.navSubPage = 'edit'; next();},
+app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
     require('./pages/instructorFileEditor/instructorFileEditor'),
 ]);
+app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/assessment/:assessment_id/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
 
 
 app.use('/pl/course_instance/:course_instance_id/instructor/assessment_instance/:assessment_instance_id', [
@@ -337,36 +361,57 @@ app.use('/pl/course_instance/:course_instance_id/instructor/assessment_instance/
     require('./pages/shared/floatFormatters'),
     require('./pages/instructorAssessmentInstance/instructorAssessmentInstance'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/edit', [
-    function(req, res, next) {res.locals.navbarType = 'instructor'; res.locals.navPage = 'question'; res.locals.navSubPage = 'edit'; next();},
-    require('./middlewares/selectAndAuthzInstructorQuestion'),
-    require('./pages/instructorFileEditor/instructorFileEditor'),
-]);
+
+// single question
+
 app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id', [
     require('./middlewares/selectAndAuthzInstructorQuestion'),
+]);
+app.use(/^(\/pl\/course_instance\/[0-9]+\/instructor\/question\/[0-9]+)\/?$/, (req, res, _next) => {
+    res.redirect(`${req.params[0]}/preview`);
+});
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id', function(req, res, next) {res.locals.navPage = 'question'; next();});
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorQuestionSettings/instructorQuestionSettings'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/preview', [
+    function(req, res, next) {res.locals.navSubPage = 'preview'; next();},
+    require('./pages/shared/floatFormatters'),
+    require('./pages/instructorQuestionPreview/instructorQuestionPreview'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/statistics', [
+    function(req, res, next) {res.locals.navSubPage = 'statistics'; next();},
     require('./pages/shared/assessmentStatDescriptions'),
     require('./pages/shared/floatFormatters'),
-    require('./pages/instructorQuestion/instructorQuestion'),
+    require('./pages/instructorQuestionStatistics/instructorQuestionStatistics'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/gradebook', require('./pages/instructorGradebook/instructorGradebook'));
-app.use('/pl/course_instance/:course_instance_id/instructor/questions', require('./pages/instructorQuestions/instructorQuestions'));
-app.use('/pl/course_instance/:course_instance_id/instructor/issues', require('./pages/instructorIssues/instructorIssues'));
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
+    require('./pages/instructorFileEditor/instructorFileEditor'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/question/:question_id/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
+
 app.use('/pl/course_instance/:course_instance_id/instructor/grading_job', require('./pages/instructorGradingJob/instructorGradingJob'));
-app.use('/pl/course_instance/:course_instance_id/instructor/syncs', require('./pages/courseSyncs/courseSyncs'));
 app.use('/pl/course_instance/:course_instance_id/instructor/jobSequence', require('./pages/instructorJobSequence/instructorJobSequence'));
 app.use('/pl/course_instance/:course_instance_id/instructor/loadFromDisk', require('./pages/instructorLoadFromDisk/instructorLoadFromDisk'));
+app.use('/pl/course_instance/:course_instance_id/instructor/edit_error', require('./pages/editError/editError'));
 
 // course instance - course admin pages
 app.use('/pl/course_instance/:course_instance_id/instructor/course_admin', [
     require('./middlewares/authzCourseInstanceHasCourseView'),
 ]);
 app.use(/^(\/pl\/course_instance\/[0-9]+\/instructor\/course_admin)\/?$/, (req, res, _next) => {
-    res.redirect(`${req.params[0]}/info`);
+    res.redirect(`${req.params[0]}/instances`);
 });
 app.use('/pl/course_instance/:course_instance_id/instructor/course_admin', function(req, res, next) {res.locals.navPage = 'course_admin'; next();});
-app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/info', [
-    function(req, res, next) {res.locals.navSubPage = 'info'; next();},
-    require('./pages/instructorCourseAdminInfo/instructorCourseAdminInfo'),
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorCourseAdminSettings/instructorCourseAdminSettings'),
 ]);
 app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/access', [
     function(req, res, next) {res.locals.navSubPage = 'access'; next();},
@@ -375,6 +420,22 @@ app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/access'
 app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/sets', [
     function(req, res, next) {res.locals.navSubPage = 'sets'; next();},
     require('./pages/instructorCourseAdminSets/instructorCourseAdminSets'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/instances', [
+    function(req, res, next) {res.locals.navSubPage = 'instances'; next();},
+    require('./pages/instructorCourseAdminInstances/instructorCourseAdminInstances'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/issues', [
+    function(req, res, next) {res.locals.navSubPage = 'issues'; next();},
+    require('./pages/instructorIssues/instructorIssues'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/questions', [
+    function(req, res, next) {res.locals.navSubPage = 'questions'; next();},
+    require('./pages/instructorQuestions/instructorQuestions'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/syncs', [
+    function(req, res, next) {res.locals.navSubPage = 'syncs'; next();},
+    require('./pages/courseSyncs/courseSyncs'),
 ]);
 app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/topics', [
     function(req, res, next) {res.locals.navSubPage = 'topics'; next();},
@@ -388,29 +449,51 @@ app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/grading
     function(req, res, next) {res.locals.navSubPage = 'grading'; next();},
     require('./pages/instructorCourseAdminGrading/instructorCourseAdminGrading'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/edit', [
-    function(req, res, next) {res.locals.navSubPage = 'edit'; next();},
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
     require('./pages/instructorFileEditor/instructorFileEditor'),
 ]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/course_admin/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
 
 // course instance - instance admin pages
 app.use(/^(\/pl\/course_instance\/[0-9]+\/instructor\/instance_admin)\/?$/, (req, res, _next) => {
-    res.redirect(`${req.params[0]}/access`);
+    res.redirect(`${req.params[0]}/assessments`);
 });
 app.use('/pl/course_instance/:course_instance_id/instructor', function(req, res, next) {res.locals.navbarType = 'instructor'; next();});
 app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin', function(req, res, next) {res.locals.navPage = 'instance_admin'; next();});
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorInstanceAdminSettings/instructorInstanceAdminSettings'),
+]);
 app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/access', [
     function(req, res, next) {res.locals.navSubPage = 'access'; next();},
     require('./pages/instructorInstanceAdminAccess/instructorInstanceAdminAccess'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/assessments', [
+    function(req, res, next) {res.locals.navSubPage = 'assessments'; next();},
+    require('./pages/instructorAssessments/instructorAssessments'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/gradebook', [
+    function(req, res, next) {res.locals.navSubPage = 'gradebook'; next();},
+    require('./pages/instructorGradebook/instructorGradebook'),
 ]);
 app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/lti', [
     function(req, res, next) {res.locals.navSubPage = 'lti'; next();},
     require('./pages/instructorInstanceAdminLti/instructorInstanceAdminLti'),
 ]);
-app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/edit', [
-    function(req, res, next) {res.locals.navSubPage = 'edit'; next();},
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
     require('./pages/instructorFileEditor/instructorFileEditor'),
 ]);
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course_instance/:course_instance_id/instructor/instance_admin/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
 
 // clientFiles
 app.use('/pl/course_instance/:course_instance_id/instructor/clientFilesCourse', require('./pages/clientFilesCourse/clientFilesCourse'));
@@ -543,14 +626,59 @@ app.use('/pl/course/:course_id', require('./middlewares/authzCourse')); // set r
 app.use('/pl/course/:course_id', function(req, res, next) {res.locals.urlPrefix = '/pl/course/' + req.params.course_id; next();});
 app.use('/pl/course/:course_id', function(req, res, next) {res.locals.navbarType = 'instructor'; next();});
 app.use(/^\/pl\/course\/[0-9]+\/?$/, function(req, res, _next) {res.redirect(res.locals.urlPrefix + '/course_admin');}); // redirect plain course URL to overview page
+app.use('/pl/course/:course_id', require('./middlewares/selectOpenIssueCount'));
+
+// single question
+
+app.use('/pl/course/:course_id/question/:question_id', [
+    require('./middlewares/selectAndAuthzInstructorQuestion'),
+]);
+app.use(/^(\/pl\/course\/[0-9]+\/question\/[0-9]+)\/?$/, (req, res, _next) => {
+    res.redirect(`${req.params[0]}/preview`);
+});
+app.use('/pl/course/:course_id/question/:question_id', function(req, res, next) {res.locals.navPage = 'question'; next();});
+app.use('/pl/course/:course_id/question/:question_id/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorQuestionSettings/instructorQuestionSettings'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/preview', [
+    function(req, res, next) {res.locals.navSubPage = 'preview'; next();},
+    require('./pages/shared/floatFormatters'),
+    require('./pages/instructorQuestionPreview/instructorQuestionPreview'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/statistics', [
+    function(req, res, next) {res.locals.navSubPage = 'statistics'; next();},
+    require('./pages/shared/assessmentStatDescriptions'),
+    require('./pages/shared/floatFormatters'),
+    require('./pages/instructorQuestionStatistics/instructorQuestionStatistics'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
+    require('./pages/instructorFileEditor/instructorFileEditor'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
+
+
+
+
+
+app.use('/pl/course/:course_id/file_transfer', [
+    require('./pages/instructorFileTransfer/instructorFileTransfer'),
+]);
+
+app.use('/pl/course/:course_id/edit_error', require('./pages/editError/editError'));
 
 app.use(/^(\/pl\/course\/[0-9]+\/course_admin)\/?$/, (req, res, _next) => {
-    res.redirect(`${req.params[0]}/info`);
+    res.redirect(`${req.params[0]}/instances`);
 });
 app.use('/pl/course/:course_id/course_admin', function(req, res, next) {res.locals.navPage = 'course_admin'; next();});
-app.use('/pl/course/:course_id/course_admin/info', [
-    function(req, res, next) {res.locals.navSubPage = 'info'; next();},
-    require('./pages/instructorCourseAdminInfo/instructorCourseAdminInfo'),
+app.use('/pl/course/:course_id/course_admin/settings', [
+    function(req, res, next) {res.locals.navSubPage = 'settings'; next();},
+    require('./pages/instructorCourseAdminSettings/instructorCourseAdminSettings'),
 ]);
 app.use('/pl/course/:course_id/course_admin/access', [
     function(req, res, next) {res.locals.navSubPage = 'access'; next();},
@@ -559,6 +687,22 @@ app.use('/pl/course/:course_id/course_admin/access', [
 app.use('/pl/course/:course_id/course_admin/sets', [
     function(req, res, next) {res.locals.navSubPage = 'sets'; next();},
     require('./pages/instructorCourseAdminSets/instructorCourseAdminSets'),
+]);
+app.use('/pl/course/:course_id/course_admin/instances', [
+    function(req, res, next) {res.locals.navSubPage = 'instances'; next();},
+    require('./pages/instructorCourseAdminInstances/instructorCourseAdminInstances'),
+]);
+app.use('/pl/course/:course_id/course_admin/issues', [
+    function(req, res, next) {res.locals.navSubPage = 'issues'; next();},
+    require('./pages/instructorIssues/instructorIssues'),
+]);
+app.use('/pl/course/:course_id/course_admin/questions', [
+    function(req, res, next) {res.locals.navSubPage = 'questions'; next();},
+    require('./pages/instructorQuestions/instructorQuestions'),
+]);
+app.use('/pl/course/:course_id/course_admin/syncs', [
+    function(req, res, next) {res.locals.navSubPage = 'syncs'; next();},
+    require('./pages/courseSyncs/courseSyncs'),
 ]);
 app.use('/pl/course/:course_id/course_admin/topics', [
     function(req, res, next) {res.locals.navSubPage = 'topics'; next();},
@@ -572,14 +716,41 @@ app.use('/pl/course/:course_id/course_admin/grading', [
     function(req, res, next) {res.locals.navSubPage = 'grading'; next();},
     require('./pages/instructorCourseAdminGrading/instructorCourseAdminGrading'),
 ]);
-app.use('/pl/course/:course_id/course_admin/edit', [
-    function(req, res, next) {res.locals.navSubPage = 'edit'; next();},
+app.use('/pl/course/:course_id/course_admin/file_edit', [
+    function(req, res, next) {res.locals.navSubPage = 'file_edit'; next();},
     require('./pages/instructorFileEditor/instructorFileEditor'),
 ]);
+app.use('/pl/course/:course_id/course_admin/file_view', [
+    function(req, res, next) {res.locals.navSubPage = 'file_view'; next();},
+    require('./pages/instructorFileBrowser/instructorFileBrowser'),
+]);
+app.use('/pl/course/:course_id/course_admin/file_download', require('./pages/instructorFileDownload/instructorFileDownload'));
 
 app.use('/pl/course/:course_id/loadFromDisk', require('./pages/instructorLoadFromDisk/instructorLoadFromDisk'));
-app.use('/pl/course/:course_id/syncs', require('./pages/courseSyncs/courseSyncs'));
 app.use('/pl/course/:course_id/jobSequence', require('./pages/instructorJobSequence/instructorJobSequence'));
+
+// clientFiles
+app.use('/pl/course/:course_id/clientFilesCourse', require('./pages/clientFilesCourse/clientFilesCourse'));
+app.use('/pl/course/:course_id/question/:question_id/clientFilesQuestion', [
+    require('./middlewares/selectAndAuthzInstructorQuestion'),
+    require('./pages/clientFilesQuestion/clientFilesQuestion'),
+]);
+
+// generatedFiles
+app.use('/pl/course/:course_id/question/:question_id/generatedFilesQuestion', [
+    require('./middlewares/selectAndAuthzInstructorQuestion'),
+    require('./pages/instructorGeneratedFilesQuestion/instructorGeneratedFilesQuestion'),
+]);
+
+// legacy client file paths
+app.use('/pl/course/:course_id/question/:question_id/file', [
+    require('./middlewares/selectAndAuthzInstructorQuestion'),
+    require('./pages/legacyQuestionFile/legacyQuestionFile'),
+]);
+app.use('/pl/course/:course_id/question/:question_id/text', [
+    require('./middlewares/selectAndAuthzInstructorQuestion'),
+    require('./pages/legacyQuestionText/legacyQuestionText'),
+]);
 
 //////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////